pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sqldown"
version = "0.1.0"
description = "Bidirectional markdown ↔ SQLite conversion with column limit protection"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Mike Bailey", email = "mike@bailey.net.au"}
]
keywords = ["markdown", "sqlite", "yaml", "frontmatter", "bidirectional"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Database",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "sqlite-utils>=3.30",
    "pyyaml>=6.0",
    "pathspec>=0.11",
    "watchdog>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
sqldown = "sqldown.cli:main"

[project.urls]
Homepage = "https://github.com/mbailey/metool-packages-dev"
Repository = "https://github.com/mbailey/metool-packages-dev"
Issues = "https://github.com/mbailey/metool-packages-dev/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/sqldown"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=sqldown --cov-report=term-missing"
